Return drainability status based on outcome priority
